fix: move set functions to a separate package
